create _patches for specific module patches and move json into it
