+++ ddtrace/appsec/iast/_patch_modules.py
@@ -20,3 +20,7 @@ def patch_iast(patch_modules=IAST_PATCH):
         when_imported("hashlib")(
             _on_import_factory(module, prefix="ddtrace.appsec.iast.taint_sinks", raise_errors=False)
         )
+
+    when_imported("json")(
+        _on_import_factory("json_tainting", prefix="ddtrace.appsec.iast._patches", raise_errors=False)
+    )
